Assay confidence data pulled and analysed for a possible filter.

+++ pull.py
@@ -27,7 +27,20 @@ COLS = [
 acts = new_client.activity.filter(target_chembl_id="CHEMBL240", assay_type="B").only(
     COLS
 )
+conf = new_client.assay.filter(target_chembl_id="CHEMBL240", assay_type="B").only(
+    "assay_chembl_id", "confidence_score"
+)
 
-df = pd.DataFrame(acts)
-df.to_csv("data/CHEMBL240_activities.csv", index=False)
-print(df.shape)
+if Path("data/CHEMBL240_activities.csv").exists() == False:
+    df = pd.DataFrame(acts)
+    df.to_csv("data/CHEMBL240_activities.csv", index=False)
+    print(df.shape)
+else:
+    print("Activities skipped")
+
+if Path("data/CHEMBL240_confidence.csv").exists() == False:
+    df_c = pd.DataFrame(conf)
+    print(df_c.shape)
+    df_c.to_csv("data/CHEMBL240_confidence.csv", index=False)
+else:
+    print("Confidence skipped")
